Fix exception vector and jump instruction: \
The exception vector consists of 16 4-byte entries.

Entries 0-7 are initialized to contain a relative jump instruction that jumps the pc
to the corresponding higher entry (8-15). E.g. entry 0 should jump the pc to entry
8, entry 1 should jump the pc to entry 9, etc. This is depicted below.

Entries 6-15 are meant to be handlers for these exception types. We currently only
use an IRQ handler, irq_dispatch. The other handlers are set to a default handler
that calls panic.

exp_vector[16]:
-------------------------------------------------------------------------------
| index | offset |      exception type        	|      	    value	      |
-------------------------------------------------------------------------------
|   0	|   0    |	reset			|	jump instruction      |
|   1	|   4	 |	undefined instruction   |	jump instruction      |
|   2   |   8	 |	software interrupt      |	jump instruction      |
|   3	|   12	 |	prefetch abort          |	jump instruction      |
|   4	|   16	 |	data abort              |	jump instruction      |
|   5	|   20	 |	**reserved**            |	jump instruction      |
|   6	|   24	 |	IRQ                     |	jump instruction      |
|   7	|   28	 |	FIQ                     |	jump instruction      |
-------------------------------------------------------------------------------
|   8	|   32	 |	reset			|	defexp_handler        |
|   9	|   36	 |	undefined instruction   |	defexp_handler        |
|   10	|   40	 |	software interrupt      |	defexp_handler        |
|   11	|   44	 |	prefetch abort          |	defexp_handler        |
|   12	|   48	 |	data abort              |	defexp_handler        |
|   13	|   52	 | 	**reserved**            |	defexp_handler        |
|   14	|   56	 |	IRQ                     |	irq_dispatch          |
|   15	|   60	 |	FIQ                     |	defexp_handler        |
-------------------------------------------------------------------------------

Previously, the jump instruction was set to jump the pc by 24 bytes and irq_dispatch
was written into the exception vector at an offset of 0x30 (48 bytes). Notice,
however, that the intended exception vector slot for irq_dispatch is actually at an
offset of 56 bytes (0x38). Furthermore, to jump from a lower index in the vector
(0-7) to the corresponding higher index, the pc must be moved by 32 bytes, not 24.

This commit fixes this up so that the exp_vector looks like the table shown above.
The jump instruction now jumps the pc by 32 bytes and irq_dispatch is now installed
at an offset of 56 bytes in exp_vector.